Shortcuts::setMetaFromRecord(): remove defensive programming.
"Fail early"; the way we use this meta stuff is by joining "model"
tables with the meta table, so the *_overwrite etc... columns will
always be present, even more so: they _should_ be present, so being
silent about them missing is probably not a good idea.  Further more, it
distracts from reading the code a lot... so this change removes all
these isset() conditionals.